+++ src/main/java/com/bitbank/responses/JwtResponse.java
@@ -6,13 +6,21 @@ public class JwtResponse {
 
 	@JsonProperty("access_token")
 	private String accessToken;
+	
+	@JsonProperty("refresh_token")
+	private String refreshToken;
 
-	public JwtResponse(String token) {
+	public JwtResponse(String token, String refreshToken) {
 		this.accessToken = token;
+		this.refreshToken = refreshToken;
 	}
 
 	public String getAccessToken() {
 		return this.accessToken;
+	}
+	
+	public String getRefreshToken() {
+		return this.refreshToken;
 	}
 
 }