+++ src/main/java/com/bitbank/services/RefreshTokensService.java
@@ -3,6 +3,7 @@ package com.bitbank.services;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.webjars.NotFoundException;
 
@@ -10,6 +11,7 @@ import com.bitbank.entities.RefreshTokensEntity;
 import com.bitbank.entities.UsersEntity;
 import com.bitbank.repositories.RefreshTokensRepository;
 import com.bitbank.repositories.UsersRepository;
+import com.bitbank.utils.AuthenticationUtils;
 import com.bitbank.utils.JwtUtils;
 import com.bitbank.utils.TimeSource;
 
@@ -26,6 +28,9 @@ public class RefreshTokensService {
 
 	@Autowired
 	private JwtUtils jwtUtils;
+	
+	@Autowired
+	private AuthenticationUtils authenticationUtils;
 
 	public Optional<RefreshTokensEntity> findByToken(String token) {
 		return refreshTokensRepository.findByToken(token);
@@ -37,9 +42,11 @@ public class RefreshTokensService {
 	 * @param userId
 	 * @return
 	 */
-	public RefreshTokensEntity createRefreshToken(Long userId) {
+	public RefreshTokensEntity createRefreshToken(Authentication authentication) {
+		
+		UserDetailsImpl userDetails = authenticationUtils.getPrincipal(authentication);
 
-		Optional<UsersEntity> optionalUsersEntity = usersRepository.findById(userId);
+		Optional<UsersEntity> optionalUsersEntity = usersRepository.findById(userDetails.getId());
 
 		if (optionalUsersEntity.isEmpty()) {
 			throw new NotFoundException("User not found");